services/ifcb_zip_storage/requirements.txt
pyifcb @ git+https://github.com/joefutrelle/pyifcb.git@v1.2.1
amplify-storage-utils[s3] @ git+https://github.com/WHOIGit/amplify-storage-utils.git@v1.5.1
PyYAML
